+++ src/platform_impl/web/mod.rs
@@ -1,5 +1,5 @@
 // Brief introduction to the internals of the web backend:
-// Currently, the web backend supports both wasm-bindgen and stdweb as methods of binding to the
+// The web backend used to support both wasm-bindgen and stdweb as methods of binding to the
 // environment. Because they are both supporting the same underlying APIs, the actual web bindings
 // are cordoned off into backend abstractions, which present the thinnest unifying layer possible.
 //
@@ -17,25 +17,14 @@
 // incoming events (from the registered handlers) and ensuring they are passed to the user in a
 // compliant way.
 
-// Silence warnings from use of deprecated stdweb backend
-#![allow(deprecated)]
-
 mod device;
 mod error;
 mod event_loop;
 mod monitor;
 mod window;
 
-#[cfg(feature = "web-sys")]
 #[path = "web_sys/mod.rs"]
 mod backend;
-
-#[cfg(feature = "stdweb")]
-#[path = "stdweb/mod.rs"]
-mod backend;
-
-#[cfg(not(any(feature = "web-sys", feature = "stdweb")))]
-compile_error!("Please select a feature to build for web: `web-sys`, `stdweb`");
 
 pub use self::device::Id as DeviceId;
 pub use self::error::OsError;